Add files via upload
Slightly updated for assignment 2 convenience

+++ Source files/TAInt.cpp
@@ -16,7 +16,7 @@ void TAInt::set(int x)
 {
 	value = x;
 }
-int TAInt::getValue()
+double TAInt::getValue()
 {
 	return value;
 }
